fix: modal setting
