+++ src/components/modals/AddSeriesModal/config.ts
@@ -1,15 +1,15 @@
-export const paymentMethodOptions = [
-  {
-    id: "bank_transfer",
-    label: "Transferencia Bancaria",
-    value: "bank_transfer",
-  },
-  { id: "credit_card", label: "Tarjeta de Crédito", value: "credit_card" },
-  { id: "cash", label: "Efectivo", value: "cash" },
-];
+  export const paymentMethodOptions = [
+    {
+      id: "bank_transfer",
+      label: "Transferencia Bancaria",
+      value: "bank_transfer",
+    },
+    { id: "credit_card", label: "Tarjeta de Crédito", value: "credit_card" },
+    { id: "cash", label: "Efectivo", value: "cash" },
+  ];
 
-export const frequencyOptions = [
-  { id: "monthly", label: "Mensual", value: "monthly" },
-  { id: "quarterly", label: "Trimestral", value: "quarterly" },
-  { id: "annually", label: "Anual", value: "annually" },
-];
+  export const frequencyOptions = [
+    { id: "monthly", label: "Mensual", value: "monthly" },
+    { id: "quarterly", label: "Trimestral", value: "quarterly" },
+    { id: "annually", label: "Anual", value: "annually" },
+  ];
