+++ src/components/modals/AddSeriesModal/type.ts
@@ -1,0 +1,24 @@
+export interface FormValues {
+    field1: number;
+    field2: number;
+  }
+  
+  export interface Option {
+    id: string;
+    label: string;
+    value: string;
+  }
+  
+  export interface AddSeriesModalProps {
+    title: string;
+    handleClose: () => void;
+    onSubmit: () => void;
+    buttonText: string;
+    secondButtonText: string;
+    portalId?: string;
+    formValues: FormValues;
+    paymentMethodOptions: Option[];
+    frequencyOptions: Option[];
+  }
+  
+
